renamed PseudoMaterials.py for PEP8
